backend: moved star system generation validation code in generateStarSystem
