+++ backend/src/lib/star-systems/generateStarSystem.ts
@@ -2,8 +2,11 @@ import type { NewStarSystem, StarSystemGenerationSettings } from "#lib/db/star-s
 import { mulberry32Prng } from "../rng/mulberry32prng.ts"
 import { BodyType } from "#lib/star-systems/BodyType.ts"
 import { createRng, type Rng } from "#lib/rng/rng.ts"
+import { Result } from "@guillaume-docquier/tools-ts"
+import type { StarSystemGenerationSettingsDto } from "#api/games/games.controller.ts"
+import type { IntegerRange, PercentageRange } from "#lib/Range.ts"
 
-export const MAX_ORBITS = 6
+const MAX_ORBITS = 6
 
 const FIRST_ORBIT_SECTOR_COUNT = 2
 const MOVEMENT_EDGE_WEIGHT = 1
@@ -285,4 +288,65 @@ function randomUuid(rng: Rng): string {
 
 function toTitleCase(value: string): string {
   return value.charAt(0) + value.slice(1).toLowerCase()
+}
+
+export function validateStarSystemGenerationSettings(settings: Readonly<StarSystemGenerationSettingsDto>): Result<true, string> {
+  const rangeChecks: Array<[string, PercentageRange | IntegerRange]> = [
+    ["planetDensity", settings.planetDensity],
+    ["nbPlanets", settings.nbPlanets],
+    ["nbMoonsPerPlanet", settings.nbMoonsPerPlanet],
+    ["nbAsteroidBelts", settings.nbAsteroidBelts],
+    ["nbAsteroidsPerSector", settings.nbAsteroidsPerSector],
+  ]
+
+  for (const [rangeName, range] of rangeChecks) {
+    if (!Number.isFinite(range.min) || !Number.isFinite(range.max)) {
+      return Result.Failure(`${rangeName} must have finite bounds`)
+    }
+
+    if (range.min > range.max) {
+      return Result.Failure(`${rangeName} min must be less than or equal to max`)
+    }
+  }
+
+  if (settings.planetDensity.min < 0 || settings.planetDensity.max > 1) {
+    return Result.Failure("planetDensity must stay between 0 and 1")
+  }
+
+  const integerRanges: Array<[string, IntegerRange]> = [
+    ["nbPlanets", settings.nbPlanets],
+    ["nbMoonsPerPlanet", settings.nbMoonsPerPlanet],
+    ["nbAsteroidBelts", settings.nbAsteroidBelts],
+    ["nbAsteroidsPerSector", settings.nbAsteroidsPerSector],
+  ]
+
+  for (const [rangeName, range] of integerRanges) {
+    if (!Number.isInteger(range.min) || !Number.isInteger(range.max)) {
+      return Result.Failure(`${rangeName} must use integer bounds`)
+    }
+
+    if (range.min < 0) {
+      return Result.Failure(`${rangeName} must be greater than or equal to 0`)
+    }
+  }
+
+  if (settings.seed !== undefined && (!Number.isInteger(settings.seed) || settings.seed < 0 || settings.seed > 2 ** 32 - 1)) {
+    return Result.Failure("seed must be an unsigned 32-bit integer")
+  }
+
+  if (settings.nbAsteroidBelts.max > MAX_ORBITS) {
+    return Result.Failure(`nbAsteroidBelts cannot be greater than ${MAX_ORBITS}`)
+  }
+
+  const maxNonBeltSectorCount = Array.from({ length: MAX_ORBITS }, (_, index) => 2 ** (index + 1))
+    .toSorted((sectorCountA, sectorCountB) => sectorCountB - sectorCountA)
+    .slice(0, MAX_ORBITS - settings.nbAsteroidBelts.max)
+    .reduce((total, sectorCount) => total + sectorCount, 0)
+  const maxPlanetCapacity = Math.floor(maxNonBeltSectorCount * settings.planetDensity.min)
+
+  if (maxPlanetCapacity < settings.nbPlanets.max) {
+    return Result.Failure(`settings cannot generate the requested Planets within ${MAX_ORBITS} orbits`)
+  }
+
+  return Result.Success(true)
 }
